Moved timers to separate script_timer.h/.cpp
